Update StoryRouter.js
when there is some exception during initialization, componentWillMount is called without componentWillMount being called first, which leads to the case when unlisten is not defined. This change allows me to see real exception, not 'this.unlisten is not a function' which says nothing about my exception. I am not sure if it is valid that willUnmount is called before componentDidMount, but this change just fixes the issue.